requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Dimensionality Reduction
umap-learn>=0.5.3
scikit-learn-extra>=0.2.0

# Clustering
hdbscan>=0.8.29
scikit-learn>=1.3.0

# Search and Similarity
faiss-cpu>=1.7.4
annoy>=1.17.3
scann>=1.2.7

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Text Processing
nltk>=3.8.1
spacy>=3.6.0
textblob>=0.17.1

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0
click>=8.1.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: GPU support
# faiss-gpu>=1.7.4  # Uncomment for GPU acceleration 